requirements.txt
Click==7.0
dill==0.2.9
Flask==1.0.3
Flask-SocketIO==4.1.0
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
multiprocess==0.70.7
pkg-resources==0.0.0
psycopg2==2.8.3
pyserial==3.4
python-engineio==3.8.1
python-socketio==4.1.0
RPi.GPIO==0.6.5
six==1.12.0
smbus2==0.2.3
Werkzeug==0.15.4
WTForms==2.2.1
